@@ -14,36 +14,42 @@
   ],
   "homepage": "https://drizzer14.github.io/fnts",
   "repository": "https://github.com/drizzer14/fnts",
+  "type": "module",
+  "module": "index.js",
+  "types": "index.d.ts",
+  "sideEffects": false,
+  "engines": {
+    "node": ">=20.13.1",
+    "npm": ">=10.5.2"
+  },
   "scripts": {
     "prebuild": "rimraf lib",
-    "copy:dts": "copyfiles --up 1 --all \"./src/**/*.d.ts\" lib",
-    "build:dts": "tsc -p tsconfig.declaration.json",
-    "build:js": "tsc -p tsconfig.build.json",
-    "build": "npm-run-all -p copy:dts build:dts build:js",
+    "build:dts": "copyfiles --up 1 --all \"./src/**/*.d.ts\" lib",
+    "build:js": "tsc --build tsconfig.build.json",
+    "postbuild:js": "tsc-alias -p tsconfig.build.json",
+    "build": "npm-run-all -p build:dts build:js",
     "lint": "tsc -p tsconfig.ci.json",
     "test": "jest",
-    "typetest": "tsc -p tsconfig.test.json",
-    "test:coverage": "jest --coverage",
-    "ci": "npm-run-all -s build lint test typetest",
-    "bump": "standard-version",
+    "typetest": "tsc -p tsconfig.spec.json",
+    "ci": "npm-run-all -s build -p lint test typetest",
+    "bump": "commit-and-tag-version",
     "prepublish": "node scripts/prepublish.js",
     "publish": "yarn publish lib --non-interactive",
     "release": "npm-run-all -s build publish"
   },
   "devDependencies": {
-    "@babel/core": "7.12.17",
-    "@types/jest": "26.0.20",
-    "babel-jest": "26.6.3",
+    "@types/jest": "29.5.12",
     "copyfiles": "2.4.1",
-    "jest": "26.6.3",
+    "jest": "29.7.0",
     "npm-run-all": "4.1.5",
     "rimraf": "3.0.2",
-    "standard-version": "9.1.1",
-    "ts-jest": "26.5.2",
-    "typescript": "4.9.3"
+    "commit-and-tag-version": "12.2.0",
+    "ts-jest": "29.1.5",
+    "tsc-alias": "1.8.10",
+    "typescript": "5.3.3"
   },
   "peerDependencies": {
-    "typescript": ">=4.7"
+    "typescript": ">=5.0"
   },
   "peerDependenciesMeta": {
     "typescript": {
