@@ -1,18 +1,20 @@
-module.exports = {
+export default {
   testEnvironment: 'node',
 	testRegex: '.*\\.spec.ts$',
 	preset: 'ts-jest',
 	rootDir: './',
 	roots: ['<rootDir>/src', '<rootDir>/tests'],
-	globals: {
-  	'ts-jest': {
-  		tsconfig: '<rootDir>/tsconfig.spec.json',
-	  }
-	},
 	transform: {
-		"\\.ts$": "ts-jest"
+		"\\.ts$": [
+			"ts-jest",
+			{
+				useESM: true,
+				tsconfig: 'tsconfig.spec.json',
+			}
+		]
 	},
   clearMocks: true,
+	extensionsToTreatAsEsm: ['.ts'],
 	moduleFileExtensions: ['ts', 'js'],
 	moduleDirectories: ['node_modules', '**/node_modules', '**/src'],
 	testPathIgnorePatterns: ['<rootDir>/lib/', '<rootDir>/node_modules/'],
