@@ -1,10 +1,10 @@
-export type { Map } from './map'
-export type { Last } from './last'
-export type { Trim } from './trim'
-export type { Slice } from './slice'
-export type { Tuple } from './tuple'
-export type { Effect } from './effect'
-export type { Gradual } from './gradual'
-export type { Unshift } from './unshift'
-export type { Flatten } from './flatten'
-export type { UnaryFunction, VariadicFunction } from './function'
+export type * from './map'
+export type * from './last'
+export type * from './trim'
+export type * from './slice'
+export type * from './tuple'
+export type * from './effect'
+export type * from './gradual'
+export type * from './unshift'
+export type * from './flatten'
+export type * from './function'
